determine current show + new logic of selecting programs
only the latest old program, the current and all future programs are
displayed. also an indicator for the current show has been added (this
is prepared to be replaced with a custom background color, however this
behaves strange and I couldn’t fix, so I commented it out for now.
maybe someone can resolve it)