Changed detection of current & future programs
This is a little more robust. In the previous implementation, a single strange entry would cause all programs to be listed because none of them were evicted from the programPlan. Now, only valid programs are added to the programPlan.

For example, right now the Google Calendar seems to return strange results, which would cause the app to display basically the entire calendar.